feat: add find functions
