+++ src/components/templateGenerator/saveSelectedPoet.js
@@ -1,0 +1,6 @@
+function savePoet() {
+  localStorage.clear();
+  localStorage.setItem('userSelect', this.innerHTML);
+}
+
+export default savePoet;
