+++ src/findPoets.js
@@ -1,4 +1,5 @@
 import data from './data/poets.json';
+import savePoet from './components/templateGenerator/saveSelectedPoet';
 
 function getDOMPoets() {
   const poetItems = document.createElement('div');
@@ -15,6 +16,7 @@ function getDOMPoets() {
         <p><img src=${item.photo} alt=${item.name} /></p>
         <figcaption><a href="./poet.html">${item.name}<a></figcaption>
       </figure>`;
+    poetItem.querySelector('a[href]').addEventListener('click', savePoet);
     poetItems.appendChild(poetItem);
   });
   if (document.getElementById('poets')) {
@@ -57,6 +59,7 @@ function getSearchPoets() {
             <p><img src=${item.photo} alt=${item.name} /></p>
             <figcaption><a href="./poet.html">${item.name}<a></figcaption>
           </figure>`;
+        searchItem.querySelector('a[href]').addEventListener('click', savePoet);
         searchItems.appendChild(searchItem);
       });
       document.getElementById('poets').appendChild(searchItems);
@@ -81,6 +84,7 @@ function getSearchPoets() {
             <p><img src=${item.photo} alt=${item.name} /></p>
             <figcaption><a href="./poet.html">${item.name}<a></figcaption>
           </figure>`;
+        searchItem.querySelector('a[href]').addEventListener('click', savePoet);
         searchItems.appendChild(searchItem);
       });
       document.getElementById('poets').appendChild(searchItems);
@@ -106,6 +110,7 @@ function getSearchPoets() {
             <p><img src=${item.photo} alt=${item.name} /></p>
             <figcaption><a href="./poet.html">${item.name}<a></figcaption>
           </figure>`;
+        searchItem.querySelector('a[href]').addEventListener('click', savePoet);
         searchItems.appendChild(searchItem);
       });
       document.getElementById('poets').appendChild(searchItems);
