+++ src/index.html
@@ -97,6 +97,5 @@
           </ul>
         </section>
       </main>
-  <script src="../build/bundle.js"></script>
 </body>
 </html>
