+++ src/poets.js
@@ -1,4 +1,5 @@
 import { addEventList, getIdPoets } from './findPoets';
+import './components/templateGenerator/saveSelectedPoet';
 
 getIdPoets();
 addEventList();
